requirements.txt
asgiref==3.3.2
attrs==19.3.0
bcrypt==4.0.1
beautifulsoup4==4.11.1
cached-property==1.5.2
certifi==2019.9.11
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
coverage==6.4.2
cryptography==38.0.3
defusedxml==0.7.1
distro==1.8.0
Django==3.2
django-cors-headers==3.13.0
django-debug-toolbar==2.2
django-filter==21.1
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
docker==6.0.1
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
Faker==12.0.1
gunicorn==20.0.4
idna==2.8
importlib-metadata==1.5.0
iniconfig==1.1.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==3.2.0
MarkupSafe==2.1.1
mixer==7.1.2
more-itertools==8.2.0
nose==1.3.7
oauthlib==3.2.2
packaging==20.1
paramiko==3.0.0
Pillow==9.0.0
pluggy==0.13.1
postgres==4.0
psycopg2-binary==2.8.6
psycopg2-pool==1.1
py==1.11.0
pycparser==2.21
PyJWT==2.1.0
PyNaCl==1.5.0
pyparsing==2.4.6
pyrsistent==0.19.3
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
python-dateutil==2.8.2
python-dotenv==0.20.0
python3-openid==3.2.0
pytz==2020.1
PyYAML==5.4.1
requests==2.26.0
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sorl-thumbnail==12.7.0
soupsieve==2.3.2.post1
sqlparse==0.3.1
texttable==1.6.7
toml==0.10.2
typing_extensions==4.3.0
uritemplate==4.1.1
urllib3==1.26.14
wcwidth==0.1.8
webcolors==1.11.1
websocket-client==0.59.0
zipp==2.2.0
